requirements.txt
streamlit
pillow
numpy
plotly
ultralytics